feat(ui): create rare evo credential card
