+++ src/ui/components/CredentialCardTemplate/CredentialCardTemplate.test.tsx
@@ -1,11 +1,37 @@
-import { act, fireEvent, render, waitFor } from "@testing-library/react";
+import { act, fireEvent, render } from "@testing-library/react";
 import { Provider } from "react-redux";
-import { store } from "../../../store";
 import { CredentialCardTemplate } from "./CredentialCardTemplate";
+import { store } from "../../../store";
 import { shortCredsFix } from "../../__fixtures__/shortCredsFix";
+import { formatShortDate } from "../../utils/formatters";
 
-describe("CredentialCardTemplate", () => {
-  it("renders Credential Card Template", async () => {
+describe("Credential Card Template", () => {
+  it("Renders Rare Evo Card Template", async () => {
+    const handleShowCardDetails = jest.fn();
+    const { getByText, getByTestId, getByAltText } = render(
+      <Provider store={store}>
+        <CredentialCardTemplate
+          name="name"
+          index={0}
+          cardData={shortCredsFix[5]}
+          isActive={true}
+          onHandleShowCardDetails={() => handleShowCardDetails(0)}
+        />
+      </Provider>
+    );
+    const card = getByTestId("rare-card-template-name-index-0");
+    expect(getByText(shortCredsFix[5].credentialType)).toBeInTheDocument();
+    expect(
+      getByText(formatShortDate(shortCredsFix[5].issuanceDate))
+    ).toBeInTheDocument();
+    expect(getByAltText(/card-logo/i)).toBeInTheDocument();
+    act(() => {
+      fireEvent.click(card);
+    });
+    expect(handleShowCardDetails).toBeCalledTimes(1);
+  });
+
+  it("Renders Keri Card Template", async () => {
     const handleShowCardDetails = jest.fn();
     const { getByText, getByTestId, getByAltText } = render(
       <Provider store={store}>
@@ -18,7 +44,8 @@ describe("CredentialCardTemplate", () => {
         />
       </Provider>
     );
-    const card = getByTestId("cred-card-template-name-index-0");
+
+    const card = getByTestId("keri-card-template-name-index-0");
     expect(getByText("Qualified vLEI Issuer Credential")).toBeInTheDocument();
     expect(getByText("22/01/2024")).toBeInTheDocument();
     expect(getByAltText(/card-logo/i)).toBeInTheDocument();
@@ -26,49 +53,5 @@ describe("CredentialCardTemplate", () => {
       fireEvent.click(card);
     });
     expect(handleShowCardDetails).toBeCalledTimes(1);
-  });
-
-  it("Click pending card", async () => {
-    const handleShowCardDetails = jest.fn();
-    const { getByTestId } = render(
-      <Provider store={store}>
-        <CredentialCardTemplate
-          name="name"
-          index={0}
-          cardData={shortCredsFix[4]}
-          isActive={true}
-          onHandleShowCardDetails={() => handleShowCardDetails(0)}
-        />
-      </Provider>
-    );
-    const card = getByTestId("cred-card-template-name-index-0");
-
-    act(() => {
-      fireEvent.click(card);
-    });
-
-    expect(handleShowCardDetails).toBeCalledTimes(0);
-
-    await waitFor(() => {
-      expect(getByTestId("alert-confirm")).toBeInTheDocument();
-    });
-  });
-
-  it("In active card status", async () => {
-    const handleShowCardDetails = jest.fn();
-    const { getByTestId } = render(
-      <Provider store={store}>
-        <CredentialCardTemplate
-          name="name"
-          index={0}
-          cardData={shortCredsFix[4]}
-          isActive={false}
-          onHandleShowCardDetails={() => handleShowCardDetails(0)}
-        />
-      </Provider>
-    );
-    const card = getByTestId("cred-card-template-name-index-0");
-
-    expect(card.classList.contains("active")).toBe(false);
   });
 });
