+++ src/ui/components/CredentialCardTemplate/components/KeriCardTemplate/KeriCardTemplate.scss
@@ -1,4 +1,4 @@
-.cred-card-template {
+.keri-card-template {
   height: calc((100vw - 2.5rem) * (200 / 350));
   width: 100%;
   padding: 1.25rem 1.56rem;
@@ -27,7 +27,7 @@
     box-shadow: 0px -5px 5px rgba(54, 60, 74, 0.25);
   }
 
-  .cred-card-template-inner {
+  .keri-card-template-inner {
     height: 100%;
     display: flex;
     flex-direction: column;
@@ -134,7 +134,7 @@
       margin-top: -8.8rem;
     }
 
-    .cred-card-template-inner {
+    .keri-card-template-inner {
       &.pending {
         .card-header {
           .card-logo {
