+++ src/ui/components/CredentialCardTemplate/components/KeriCardTemplate/index.ts
@@ -1,0 +1,1 @@
+export * from "./KeriCardTemplate";
