+++ src/ui/components/CredentialCardTemplate/components/RareEvoCardTemplate/RareEvoCardTemplate.scss
@@ -1,0 +1,199 @@
+.rare-card-template {
+  height: calc((100vw - 2.5rem) * (200 / 350));
+  width: 100%;
+  padding: 1.25rem 1.56rem;
+  border-radius: 1rem;
+  font-size: 0.875rem;
+  line-height: 1.375rem;
+  overflow: hidden;
+  text-shadow: 0px 2px 4px rgba(12, 64, 90, 0.5);
+
+  &.active:not(:first-child).picked-card {
+    // @TODO - foconnor: We should avoid this - animating with top.
+    position: absolute;
+    top: 0rem !important;
+    margin-top: 0;
+    transition: top 0.25s;
+    box-shadow: none;
+  }
+
+  &.active.not-picked {
+    transform: translate3d(0, 100vh, 0);
+    transition: 0.5s ease-out;
+  }
+
+  &:not(:first-child) {
+    margin-top: -10rem;
+    box-shadow: 0px -5px 5px rgba(54, 60, 74, 0.25);
+  }
+
+  .rare-card-template-inner {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    color: var(--ion-color-light);
+
+    &.pending .card-footer-column-label {
+      color: white;
+    }
+
+    .card-header {
+      display: flex;
+      justify-content: space-between;
+      font-weight: bold;
+
+      .card-logo {
+        display: flex;
+        flex-direction: row;
+        line-height: normal;
+
+        img {
+          margin-right: 0.5rem;
+          max-height: 1.375rem;
+        }
+      }
+    }
+
+    .card-footer {
+      display: flex;
+      flex-direction: row;
+      justify-content: flex-end;
+
+      .card-footer-column {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+
+        &:first-child {
+          .card-footer-column-label,
+          .card-footer-column-value {
+            text-align: left;
+            width: calc(var(--scale-w) * 40);
+          }
+        }
+
+        &:last-child {
+          .card-footer-column-label,
+          .card-footer-column-value {
+            text-align: right;
+            width: calc(var(--scale-w) * 30);
+          }
+        }
+
+        .card-footer-column-label {
+          font-size: 0.625rem;
+          font-weight: 700;
+          text-transform: uppercase;
+        }
+
+        .card-footer-column-value {
+          font-weight: 700;
+          font-size: 0.875rem;
+        }
+      }
+    }
+
+    &.pending {
+      .card-footer .card-footer-column .card-footer-column-value {
+        border-radius: 0.25rem;
+        background: var(--ion-color-medium-grey);
+      }
+
+      ion-chip {
+        max-height: 1.625rem;
+        font-weight: 400;
+        font-size: 0.75rem;
+        background: var(--ion-color-medium-grey);
+        border-radius: 0.25rem;
+        margin: 0;
+        padding: 0.375rem 0.625rem;
+
+        ion-icon {
+          width: 0.875rem;
+          height: 0.875rem;
+          margin: 0;
+        }
+
+        span {
+          text-transform: capitalize;
+          margin-left: 0.31rem;
+        }
+      }
+    }
+  }
+
+  @media screen and (min-width: 250px) and (max-width: 370px) {
+    height: calc((100vw - 1.8rem) * (200 / 350));
+    padding: 1.06rem 1.326rem;
+    border-radius: 0.85rem;
+    font-size: 0.75rem;
+    line-height: 1.169rem;
+
+    &:not(:first-child) {
+      margin-top: -8.8rem;
+    }
+
+    .rare-card-template-inner {
+      .card-footer {
+        .card-footer-column {
+          .card-footer-column-label {
+            font-size: 0.625rem;
+          }
+
+          .card-footer-column-value {
+            font-size: 0.75rem;
+          }
+        }
+      }
+
+      &.pending {
+        .card-header {
+          .card-logo {
+            img {
+              margin-right: 0.375rem;
+              max-height: 1.03125rem;
+            }
+          }
+        }
+
+        .card-footer {
+          .card-footer-column {
+            .card-footer-column-label {
+              font-size: 0.46875rem;
+            }
+
+            .card-footer-column-value {
+              border-radius: 0.1875rem;
+            }
+
+            ion-chip {
+              max-height: 1.21875rem;
+              font-size: 0.5625rem;
+              border-radius: 0.1875rem;
+              padding: 0.28125rem 0.46875rem;
+
+              ion-icon {
+                width: 0.65625rem;
+                height: 0.65625rem;
+                margin: 0;
+              }
+
+              span {
+                margin-left: 0.2325rem;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  @media screen and (min-width: 480px) {
+    height: 15.375rem;
+
+    &:not(:first-child) {
+      margin-top: -12rem;
+    }
+  }
+}
