+++ src/ui/components/CredentialCardTemplate/components/RareEvoCardTemplate/index.ts
@@ -1,0 +1,1 @@
+export * from "./RareEvoCardTemplate";
