+++ src/ui/components/CredentialCardTemplate/components/index.ts
@@ -1,0 +1,2 @@
+export * from "./KeriCardTemplate";
+export * from "./RareEvoCardTemplate";
